allow saving a research link

+++ base/application/controllers/admin_panel/results/Research.php
@@ -82,7 +82,7 @@ class Research extends MY_Controller
             show_404();
         }
 
-        $this->form_validation->set_rules('description', 'Short Description', 'required|max_length[85]');
+        $this->form_validation->set_rules('description_short', 'Short Description', 'required|max_length[85]');
         $this->form_validation->set_rules('url', 'URL', 'required|valid_url');
         $this->form_validation->set_rules('field_id', 'Field', 'required|numeric');
 
@@ -90,14 +90,17 @@ class Research extends MY_Controller
         if ($this->form_validation->run() == true) {
 
             // POST data
-            $description = $this->input->post('description');
+            $description_short = $this->input->post('description_short');
             $url = $this->input->post('url');
             $field_id = $this->input->post('field_id');
 
-            $this->research->create($description, $url, $field_id);
+            $create = $this->research->create($description_short, $url, $field_id);
 
-            // Display success flash message
-            $this->session->set_flashdata('success', 1);
+            if ($create) {
+                $this->session->set_flashdata('success', 1);
+            } else {
+                $this->session->set_flashdata('failure', 0);
+            }
 
             redirect('/admin/results/research/add');
         }
@@ -118,45 +121,70 @@ class Research extends MY_Controller
             show_404();
         }
 
-        $this->form_validation->set_rules('title', 'Title', 'required');
-        $this->form_validation->set_rules('description_short', 'Short Description', 'required|max_length[85]');
-        $this->form_validation->set_rules('field_id', 'Field', 'required|numeric');
-        $this->form_validation->set_rules('priority', 'Priority', 'required|numeric');
-        $this->form_validation->set_rules('display_url', 'Home Display');
-        $this->form_validation->set_rules('www', 'WWW', 'required|valid_url');
-        $this->form_validation->set_rules('enabled', 'Enabled', 'required|numeric');
-        $this->form_validation->set_rules('redirect', 'Redirect', 'required|numeric');
-
+        if ($this->input->post('action') == 'save') {
+            $this->form_validation->set_rules('title', 'Title', 'trim');
+            $this->form_validation->set_rules('description_short', 'Short Description', 'required|trim|max_length[85]');
+            $this->form_validation->set_rules('field_id', 'Field', 'required');
+            $this->form_validation->set_rules('display_url', 'Home Display', 'trim');
+            $this->form_validation->set_rules('url', 'URL', 'required|valid_url');
+        } else if ($this->input->post('action') == 'submit') {
+            $this->form_validation->set_rules('title', 'Title', 'required|trim');
+            $this->form_validation->set_rules('description_short', 'Short Description', 'required|trim|max_length[85]');
+            $this->form_validation->set_rules('field_id', 'Field', 'required');
+            $this->form_validation->set_rules('priority', 'Priority', 'required');
+            $this->form_validation->set_rules('display_url', 'Home Display', 'trim');
+            $this->form_validation->set_rules('url', 'URL', 'required|valid_url');
+            $this->form_validation->set_rules('enabled', 'Enabled', 'required');
+            $this->form_validation->set_rules('redirect', 'Redirect', 'required');
+        }
 
         if ($this->form_validation->run() == true) {
 
             $title = $this->input->post('title');
             $description_short = $this->input->post('description_short');
+            $url = $this->input->post('url');
             $display_url = $this->input->post('display_url');
-            $www = $this->input->post('www');
             $field_id = $this->input->post('field_id');
             $priority = $this->input->post('priority');
             $enabled = $this->input->post('enabled');
             $redirect = $this->input->post('redirect');
 
-            // create ad link
-            $this->fields->create_result_listing($title, $enabled, $description_short, $display_url, $www, $field_id, $priority, $redirect);
+            if ($this->input->post('action') == 'submit') {
 
-            // delete the research link
-            $this->delete($id);
+                // create ad link
+                $create = $this->fields->create_result_listing($title, $enabled, $description_short, $display_url, $url, $field_id, $priority, $redirect);
 
-            // Display success flash message
-            $this->session->set_flashdata('success', 1);
+                if ($create) {
+                    // delete the research link
+                    $this->delete($id);
+                    $this->session->set_flashdata('submit_success', 1);
+                } else {
+                    $this->session->set_flashdata('submit_failure', 0);
+                }
+            } elseif (($this->input->post('action') == 'save')) {
+                // save research link information
+                $update = $this->research->update($id, $title, $description_short, $url, $display_url, $field_id, $enabled, $redirect);
+
+                if ($update) {
+                    $this->session->set_flashdata('save_success', 1);
+                } else {
+                    $this->session->set_flashdata('save_failure', 0);
+                }
+            }
 
             redirect('/admin/results/research/list');
         }
 
         $research = $this->research->get($id);
 
-        $data['description'] = $research->description;
+        $data['link_title'] = $research->title;
+        $data['description_short'] = $research->description_short;
         $data['url'] = $research->url;
+        $data['display_url'] = $research->display_url;
         $data['field']['id'] = $research->field_id;
         $data['field']['name'] = $research->field;
+        $data['enabled'] = $research->enabled;
+        $data['redirect'] = $research->redirect;
 
         $prioritiesObject = $this->fields->get_links_priority($research->field_id);
         $priorities = array();
